Vectors are more DDD
